+++ lib/whois/record/parser/whois.verisign-grs.com.rb
@@ -21,6 +21,11 @@ module Whois
       #
       class WhoisVerisignGrsCom < BaseVerisign
 
+        property_supported :expires_on do
+          node("Expiration Date") { |value| Time.parse(value) }
+        end
+
+
         property_supported :registrar do
           node("Registrar") do |value|
             Whois::Record::Registrar.new(